+++ plugins/postcss-color-functional-notation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to PostCSS Color Functional Notation
 
+### 6.0.7
+
+_March 17, 2024_
+
+- Updated [`@csstools/css-color-parser`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-color-parser) to [`1.6.2`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-color-parser/CHANGELOG.md#162) (patch)
+
 ### 6.0.6
 
 _March 13, 2024_