+++ plugins/postcss-color-mix-function/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to PostCSS Color Mix Function
 
+### 2.0.12
+
+_March 17, 2024_
+
+- Updated [`@csstools/css-color-parser`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-color-parser) to [`1.6.2`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-color-parser/CHANGELOG.md#162) (patch)
+
 ### 2.0.11
 
 _March 13, 2024_